Extract components for meetups page
